Disable ability to click outside of the upload dialog
Currently, it's possible to click outside of the upload dialog. This
doesn't cancel the upload, but it makes for a weird user experience,
since the user is under the expectation that they need to keep the
window open for the dialog to complete.

Eventually, it would probably make sense to give them an action that
manually triggers a cancellation of the upload. Maybe.

Closes #5