first pass: allow metadata extraction to take place in subtasks
it can be helpful for us to break down the metadata extraction steps
into subtasks that allow us to process some of the requisite data in
parallel. we'll use a content cache to speed up the generation process.
this should also help us sidestep some rate limiting issues for large
outputs.