Initialize experimental revisions for extracting images from PDF
Extracting images from the source PDF can be useful for a number of
reasons, primarily because it can help the reader glean a better
first-pass understanding of the paper faster.

We can incorporate the images into the initial brief with inline graphs
for support, or additional use them in the chat responses, or have a
separate section for extracted relevant metadata for charts, graphs.

This may also particularly be useful in the search functionality for
retrieving relevant charts.