replace existing images with placeholder IDs in the PDF
in order to allow the LLM to work with the images in the existing PDF,
replace the existing image xrefs with place holder text-only IDs that it
can then use to insert inline images in its response. this requires us
to add some additional pre-processing to the PDF up front that might
cause a performance hit, as we need to do pdf-generation and replacement
up front.

open question: how will this be re-usable in chat? should we just tell
the LLM to use the title, like `fig x` if it wants to use it, and then
cross reference with the caption?